requirements.txt
# Data manipulation
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Hyperparameter optimization
optuna==3.4.0

# Experiment tracking and versioning
mlflow==2.9.1
dvc[s3]==3.34.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
